requirements.txt
PyYAML
boto3
termcolor
